+++ src/extension.ts
@@ -35,9 +35,8 @@ import {
   compileExplorerItems,
   checkChangedOnServer,
   compileOnly,
-  importLocalFilesToServerSideFolder,
   loadChanges,
-  importXMLFiles,
+  importArbitraryFiles,
 } from "./commands/compile";
 import { deleteExplorerItems } from "./commands/delete";
 import { exportAll, exportCurrentFile, exportDocumentsToXMLFile, exportExplorerItems } from "./commands/export";
@@ -807,6 +806,7 @@ function sendWsFolderTelemetryEvent(wsFolders: readonly vscode.WorkspaceFolder[]
       dockerCompose: !serverSide ? String(typeof conf.get("conn.docker-compose") == "object") : undefined,
       "config.conn.links": String(Object.keys(conf.get("conn.links", {})).length),
       "config.refreshClassesOnSync": !serverSide ? conf.get("refreshClassesOnSync") : undefined,
+      "config.insertStubContent": !serverSide ? conf.get("insertStubContent") : undefined,
     });
   });
 }
@@ -1142,7 +1142,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<any> {
       const file = currentFile();
       if (!file) return;
       try {
-        await loadChanges([file]);
+        await loadChanges([file], true);
       } catch (error) {
         handleError(
           error,
@@ -1409,8 +1409,14 @@ export async function activate(context: vscode.ExtensionContext): Promise<any> {
     vscode.workspace.onDidCreateFiles((e: vscode.FileCreateEvent) =>
       e.files
         // Attempt to fill in stub content for classes and routines that
-        // are not server-side files and were not created due to an export
-        .filter((f) => notIsfs(f) && isClassOrRtn(f.path) && !exportedUris.has(f.toString()))
+        // are client-side files and were not created due to an export
+        .filter(
+          (f) =>
+            notIsfs(f) &&
+            isClassOrRtn(f.path) &&
+            !exportedUris.has(f.toString()) &&
+            vscode.workspace.getConfiguration("objectscript", f).get<boolean>("insertStubContent")
+        )
         .forEach(async (uri) => {
           // Need to wait in case file was created using "Save As..."
           // because in that case the file gets created without
@@ -1646,19 +1652,6 @@ export async function activate(context: vscode.ExtensionContext): Promise<any> {
     }),
     vscode.window.registerFileDecorationProvider(fileDecorationProvider),
     vscode.workspace.onDidOpenTextDocument((doc) => !doc.isUntitled && fileDecorationProvider.emitter.fire(doc.uri)),
-    vscode.commands.registerCommand("vscode-objectscript.importLocalFilesServerSide", (wsFolderUri) => {
-      sendCommandTelemetryEvent("importLocalFilesServerSide");
-      if (
-        wsFolderUri instanceof vscode.Uri &&
-        wsFolderUri.scheme == FILESYSTEM_SCHEMA &&
-        (vscode.workspace.workspaceFolders != undefined
-          ? vscode.workspace.workspaceFolders.some((wsFolder) => wsFolder.uri.toString() == wsFolderUri.toString())
-          : false)
-      ) {
-        // wsFolderUri is an isfs workspace folder URI
-        return importLocalFilesToServerSideFolder(wsFolderUri);
-      }
-    }),
     vscode.commands.registerCommand("vscode-objectscript.modifyWsFolder", (wsFolderUri?: vscode.Uri) => {
       sendCommandTelemetryEvent("modifyWsFolder");
       modifyWsFolder(wsFolderUri);
@@ -1727,9 +1720,9 @@ export async function activate(context: vscode.ExtensionContext): Promise<any> {
       // Send telemetry events for all added folders
       sendWsFolderTelemetryEvent(e.added, true);
     }),
-    vscode.commands.registerCommand("vscode-objectscript.importXMLFiles", () => {
-      sendCommandTelemetryEvent("importXMLFiles");
-      importXMLFiles();
+    vscode.commands.registerCommand("vscode-objectscript.importFiles", () => {
+      sendCommandTelemetryEvent("importFiles");
+      importArbitraryFiles();
     }),
     vscode.commands.registerCommand("vscode-objectscript.exportToXMLFile", () => {
       sendCommandTelemetryEvent("exportToXMLFile");
@@ -1840,14 +1833,24 @@ export async function activate(context: vscode.ExtensionContext): Promise<any> {
       sendCommandTelemetryEvent("showAllClassMembers");
       if (uri instanceof vscode.Uri) showAllClassMembers(uri);
     }),
-    vscode.workspace.onDidSaveTextDocument((d) => {
+    vscode.workspace.onDidSaveTextDocument(async (d) => {
       // If the document just saved is a server-side document that needs to be updated in the UI,
       // then force VS Code to update the document's contents. This is needed if the document has
       // been changed during a save, for example by adding or changing the Storage definition.
       if (notIsfs(d.uri)) return;
       const uriString = d.uri.toString();
       if (fileSystemProvider.needsUpdate(uriString)) {
-        const activeDoc = vscode.window.activeTextEditor?.document;
+        let activeDoc = vscode.window.activeTextEditor?.document;
+        if (activeDoc?.uri.toString() != uriString) {
+          // The active text editor (if any) does not contain this document. Wait a short time and
+          // check again in case this document was saved using the "Save As..." command. In that
+          // case VS Code saves the document once with no content and then saves it again with content
+          // from the source document. During that second save our FileSystemProvider will likely
+          // change the content of the document so the header matches the new URI. We need to force
+          // VS Code to reflect that change in the editor UI.
+          await new Promise((resolve) => setTimeout(resolve, 50));
+          activeDoc = vscode.window.activeTextEditor?.document;
+        }
         if (activeDoc && !activeDoc.isDirty && !activeDoc.isClosed && activeDoc.uri.toString() == uriString) {
           // Force VS Code to refresh the file's contents in the editor tab
           vscode.commands.executeCommand("workbench.action.files.revert");